utils: Move SubProject definition to utils.universal
Since we use it at ever layer of the codebase, it doesn't make sense for
it to live in the interpreter
