+++ mesonbuild/interpreter/primitives/range.py
@@ -13,7 +13,7 @@ from ...interpreterbase import (
 )
 
 if T.TYPE_CHECKING:
-    from ...interpreterbase import SubProject
+    from ...mesonlib import SubProject
 
 class RangeHolder(MesonInterpreterObject, IterableObject):
     def __init__(self, start: int, stop: int, step: int, *, subproject: 'SubProject') -> None:
