+++ mesonbuild/interpreterbase/__init__.py
@@ -48,8 +48,6 @@ __all__ = [
 
     'InterpreterBase',
 
-    'SubProject',
-
     'TV_func',
     'TYPE_elementary',
     'TYPE_var',
@@ -81,8 +79,6 @@ from .baseobjects import (
     TYPE_nkwargs,
     TYPE_key_resolver,
     TYPE_HoldableTypes,
-
-    SubProject,
 
     HoldableTypes,
