+++ mesonbuild/modules/hotdoc.py
@@ -29,7 +29,7 @@ if T.TYPE_CHECKING:
     from ..environment import Environment
     from ..interpreter import Interpreter
     from ..interpreter.kwargs import TargetDepends
-    from ..interpreterbase import TYPE_kwargs, TYPE_var, SubProject
+    from ..interpreterbase import TYPE_kwargs, TYPE_var
 
     _T = T.TypeVar('_T')
 
@@ -387,7 +387,7 @@ class HotdocTargetHolder(_CustomTargetHolder['HotdocTarget']):
 
 
 class HotdocTarget(CustomTarget):
-    def __init__(self, name: str, subdir: str, subproject: SubProject, hotdoc_conf: File,
+    def __init__(self, name: str, subdir: str, subproject: mesonlib.SubProject, hotdoc_conf: File,
                  extra_extension_paths: T.Set[str], extra_assets: T.List[str],
                  subprojects: T.List['HotdocTarget'], environment: Environment, **kwargs: T.Any):
         super().__init__(name, subdir, subproject, environment, **kwargs, absolute_paths=True)
