+++ mesonbuild/wrap/wrap.py
@@ -32,10 +32,9 @@ from . import WrapMode
 from .. import coredata
 from ..mesonlib import (
     DirectoryLock, DirectoryLockAction, quiet_git, GIT, ProgressBar, MesonException,
-    windows_proof_rmtree, Popen_safe
+    windows_proof_rmtree, Popen_safe, SubProject,
 )
 from ..interpreterbase import FeatureNew
-from ..interpreterbase import SubProject
 from .. import mesonlib
 
 if T.TYPE_CHECKING:
